+++ src/main/java/frc/robot/subsystems/swerve/SwerveIOReal.kt
@@ -1,0 +1,82 @@
+package frc.robot.subsystems.swerve
+
+import com.ctre.phoenix6.swerve.SwerveDrivetrain
+import com.ctre.phoenix6.swerve.SwerveModule
+import com.ctre.phoenix6.swerve.SwerveRequest
+import edu.wpi.first.units.Units.*
+import edu.wpi.first.wpilibj2.command.Command
+import frc.robot.RobotContainer
+import frc.robot.constants.TunerConstants
+import frc.robot.util.Telemetry
+import java.util.function.Consumer
+import java.util.function.Supplier
+
+class SwerveIOReal() : SwerveIO {
+    private val drivetrain = TunerConstants.createDrivetrain()
+
+    private val logger: Telemetry = Telemetry(SwerveConstants.MaxSpeedConst)
+
+
+    init {
+        drivetrain.registerTelemetry(logger::telemeterize)
+    }
+
+    override fun setDefaultCommand(command: Command) {
+        drivetrain.defaultCommand = command
+    }
+
+    override fun seedFieldCentric() {
+        drivetrain.seedFieldCentric()
+    }
+
+    /* Setting up bindings for necessary control of the swerve drive platform */
+    private val fieldCentric: SwerveRequest.FieldCentric = SwerveRequest.FieldCentric()
+        .withDeadband(SwerveConstants.MaxSpeedConst * 0.1)
+        .withRotationalDeadband(SwerveConstants.MaxAngularRateConst * 0.1) // Add a 10% deadband
+        .withDriveRequestType(SwerveModule.DriveRequestType.OpenLoopVoltage) // Use open-loop control for drive motors
+
+    private val robotRelative: SwerveRequest.RobotCentric = SwerveRequest.RobotCentric()
+        .withDeadband(SwerveConstants.MaxSpeedConst * 0.1)
+        .withRotationalDeadband(SwerveConstants.MaxAngularRateConst * 0.1) // Add a 10% deadband
+        .withDriveRequestType(SwerveModule.DriveRequestType.OpenLoopVoltage) // Use open-loop control for drive motors
+
+
+    private val brake = SwerveRequest.SwerveDriveBrake()
+    private val point = SwerveRequest.PointWheelsAt()
+
+    override fun applyDriveRequest(x: Double, y: Double, twist: Double) {
+        drivetrain.applyRequest {
+            fieldCentric.withVelocityX(
+                x
+            ) //
+                // Drive forward with
+                // negative Y (forward)
+                .withVelocityY(
+                    y
+                ) // Drive
+                // left with negative X
+                // (left)
+                .withRotationalRate(
+                    twist
+                )
+        }
+    }
+
+    override fun applyRobotRelativeDriveRequest(x: Double, y: Double, twist: Double) {
+        drivetrain.applyRequest {
+            robotRelative.withVelocityX(
+                x
+            ) //
+                // Drive forward with
+                // negative Y (forward)
+                .withVelocityY(
+                    y
+                ) // Drive
+                // left with negative X
+                // (left)
+                .withRotationalRate(
+                    twist
+                )
+        }
+    }
+}
