+++ src/main/java/frc/robot/subsystems/swerve/SwerveIOSim.kt
@@ -1,0 +1,22 @@
+package frc.robot.subsystems.swerve
+
+import edu.wpi.first.wpilibj2.command.Command
+
+class SwerveIOSim() : SwerveIO {
+    override fun setDefaultCommand(command: Command) {
+        TODO("Not yet implemented")
+    }
+
+    override fun seedFieldCentric() {
+        TODO("Not yet implemented")
+    }
+
+    override fun applyDriveRequest(x: Double, y: Double, twist: Double) {
+        TODO("Not yet implemented")
+    }
+
+    override fun applyRobotRelativeDriveRequest(x: Double, y: Double, twist: Double) {
+        TODO("Not yet implemented")
+    }
+
+}
